pr 리뷰 반영

+++ src/main/java/com/jaegokeeper/hwan/item/dto/PageResponseDTO.java
@@ -12,6 +12,6 @@ public class PageResponseDTO<T>{
     private List<T> content;
     private Integer page;
     private Integer size;
-    private Long totalElements;
+    private Integer totalElements;
     private Integer totalPages;
 }
